+++ src/components/Board/RankingButton.tsx
@@ -1,0 +1,23 @@
+import React from "react";
+import styled from "styled-components";
+import { Link } from "react-router-dom";
+
+const StyledImg = styled.img<{ right?: boolean }>`
+  width: 40px;
+  height: 40px;
+  align-self: center;
+  justify-self: ${({ right }) => (right ? "right" : "center")};
+  z-index: 999;
+`;
+
+const StyledLink = StyledImg.withComponent(Link);
+
+const RankingButton = () => {
+  return (
+    <StyledLink to="/ranking">
+      <StyledImg src="/images/trophy.png" alt="ranking" />
+    </StyledLink>
+  );
+};
+
+export default RankingButton;
